+++ init.go
@@ -6,4 +6,6 @@ var (
 	_bpool = internal.NewBufferPool()
 	_cps   = new(compressors)
 	_dps   = new(decompressors).init()
+
+	JsonCodec = new(jsonCodec)
 )
